+++ packages/ts-utils/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## 0.7.0
+## 0.7.1
+
+(The 0.7.0 release was deprecated because its `validation` package was faulty.)
 
 * Initial creation and publication of this package, as an extraction and de-duplication from `@lionweb/core`, `@lionweb/utilities`, `@lionweb/class-core`, and `@lionweb/validation`.
 * Add 'mapFrom' function that maps an array to a map, using given key and value functions.